# Conflicts:
#	Src/ADOComponents/ADOComponents.vcxproj
#	Src/DXFramework/DXFramework.vcxproj
#	Src/DataInterface/Code/TelegramBuilder.cpp
#	Src/DataInterface/DataInterface.vcxproj
#	Src/DataInterface/Include/TelegramScheme.h
#	Src/EncryptionLib/EncryptionLib.vcxproj
#	Src/MESLib/Code/MESBaseADODB.cpp
#	Src/MESLib/Code/XCOM.cpp
#	Src/MESLib/Include/BaseCodec.h
#	Src/MESLib/Include/XCOM.h
#	Src/OPCOperater/OPCOperater.vcxproj
#	Tools/IPCapture/IPCapture.vcxproj

+++ Src/SocketNetwork/SimpleProtocol.cs
@@ -37,7 +37,7 @@ namespace GLEO.MES.Network
 
         public void Add(int data)
         {
-            this.Buffer.SetData(BitConverter.GetBytes((short)data));
+            this.Buffer.SetData(BitConverter.GetBytes(data));
         }
 
         public void Add(string data)
